+++ changelog.md
@@ -14,6 +14,7 @@ Codion Change Log
 - DefaultListBoxBuilder bug fixed, did not populate combo box model with initial values.
 - Controls.empty(), notEmpty() and get(index) removed, related refactoring.
 - Controls.Builder no longer adds empty Controls instances, skips leading and trailing separators, related refactoring.
+- FilterTable.Builder.startEditing(keyStroke) added.
 ### is.codion.framework.domain
 - DualValueColumnCondition now throws exception if a bound value is null.
 - SingleValueColumnCondition now throws exception if a required bound value is null.